aws-ssm-parameter/src/main/java/com/amazonaws/ssm/parameter/DeleteHandler.java
package com.amazonaws.ssm.parameter;

import software.amazon.awssdk.services.ssm.SsmClient;
import software.amazon.awssdk.services.ssm.model.ParameterNotFoundException;
import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
import software.amazon.cloudformation.proxy.Logger;
import software.amazon.cloudformation.proxy.ProgressEvent;
import software.amazon.cloudformation.proxy.OperationStatus;
import software.amazon.cloudformation.proxy.ResourceHandlerRequest;

public class DeleteHandler extends BaseHandler<CallbackContext> {
    private static final SsmClient ssmClient = SSMClientBuilder.getClient();

    @Override
    public ProgressEvent<ResourceModel, CallbackContext> handleRequest(
        final AmazonWebServicesClientProxy proxy,
        final ResourceHandlerRequest<ResourceModel> request,
        final CallbackContext callbackContext,
        final Logger logger) {

        final ResourceModel model = request.getDesiredResourceState();

        try {
            proxy.injectCredentialsAndInvokeV2(Translator.deleteParameterRequest(model), ssmClient::deleteParameter);

            return ProgressEvent.<ResourceModel, CallbackContext>builder()
                    .resourceModel(model)
                    .status(OperationStatus.SUCCESS)
                    .build();
        } catch (ParameterNotFoundException e) {
            // Parameter Store service throws this exception when the Parameter to be deleted is not found. However, this issue might also
            // occur when the first call to Lambda times out(but ends up deleting the parameter) and a second call is initiated
            // as a part of retry mechanism which ends up in an error state. So, in the interest of good customer experience,
            // we swallow this exception in CloudFormation and return a success status.
            logger.log(
                    String.format("Setting status to SUCCESS since Parameter [%s] not found.", model.getName()));

            return ProgressEvent.<ResourceModel, CallbackContext>builder()
                    .resourceModel(model)
                    .status(OperationStatus.SUCCESS)
                    .build();
        }
    }
}
